faker seeding added

+++ database/migrations/2018_12_06_130444_create_products_table.php
@@ -17,7 +17,8 @@ class CreateProductsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->text('detail');
-            $table->decimal('price');
+            // $table->decimal('price', 4,2);
+            $table->integer('price');
             $table->integer('stock');
             $table->integer('discount');
             $table->timestamps();
